+++ src/tar.h
@@ -1,0 +1,6 @@
+#ifndef __TAR_H_
+#define __TAR_H_
+
+void tar(int ac, char* av[]);
+
+#endif